+++ cvrp_ocl/problem.h
@@ -27,9 +27,9 @@ using namespace std;
 extern int g_master_problem_iteration_num;   /* 每次外循环，主问题蚁群的迭代的次数 */
 extern int g_sub_problem_iteration_num;      /* 每次外循环，子问题蚁群的迭代次数 */
 
-extern double rho;           /* parameter for evaporation */
-extern double alpha;         /* importance of trail */
-extern double beta;          /* importance of heuristic evaluate */
+extern float rho;           /* parameter for evaporation */
+extern float alpha;         /* importance of trail */
+extern float beta;          /* importance of heuristic evaluate */
 extern int ras_ranks;   /* additional parameter for rank-based version of ant system */
 extern bool sa_flag;
 
@@ -39,8 +39,8 @@ enum DistanceTypeEnum {
 };
 
 struct Point {
-    double x;
-    double y;
+    float x;
+    float y;
     int demand;     /* 每个配送点需求 */
 };
 
@@ -52,11 +52,11 @@ struct RouteCenter {
 };
 
 struct Route {
-    Route(int beg_, int end_, int load_, double dist_):beg(beg_), end(end_), load(load_), dist(dist_){}
+    Route(int beg_, int end_, int load_, float dist_):beg(beg_), end(end_), load(load_), dist(dist_){}
     int beg;                   /* route在tour中的开始位置,tour[beg] = 0(depot) */
     int end;                   /* route在tour中的结束位置,tour[end] = 0(depot) */
     int load;                  /* route load */
-    double dist;                    /* route distance */
+    float dist;                    /* route distance */
 };
 
 /*
@@ -67,7 +67,7 @@ struct Route {
 struct AntStruct {
     int  *tour;
     int  tour_size;     /* 路程中经过的点个数（depot可能出现多次） */
-    double  tour_length;     /* 车辆路程行走长度 */
+    float  tour_length;     /* 车辆路程行走长度 */
     bool      *visited;      /* mark nodes visited status */
     bool      *demand_meet_node;  /** 所有可配送的点(单次route中，目前车辆可以仍可配送的点) */
 };
@@ -80,16 +80,16 @@ struct Problem {
     char          name[LINE_BUF_LEN];      	 /* instance name */
     char          edge_weight_type[LINE_BUF_LEN];  /* selfexplanatory */
     DistanceTypeEnum dis_type;               /* 用于决定使用哪种距离方式 */
-    double        optimum;                /* optimal tour length if known, otherwise a bound */
+    float        optimum;                /* optimal tour length if known, otherwise a bound */
     int      num_node;               /* number of nodes, depot included, numd_node = 1(depot) + num of target nodes*/
     Point         *nodeptr;               /* array of structs containing coordinates of nodes */
-    double        **distance;             /* distance matrix: distance[i][j] gives distance
+    float        **distance;             /* distance matrix: distance[i][j] gives distance
                                            between node i und j */
     int      **nn_list;              /* nearest neighbor list; contains for each node i a
                                            sorted list of n_near nearest neighbors */
     int      vehicle_capacity;       /* 车辆最大装载量 */
-    double        max_distance;           /* 最大行驶距离 */
-    double        service_time;           /*  service time needed for node */
+    float        max_distance;           /* 最大行驶距离 */
+    float        service_time;           /*  service time needed for node */
     
     
     /*----- local search -----*/
@@ -107,29 +107,29 @@ struct Problem {
     AntStruct *best_so_far_ant;        /* struct that contains the best-so-far ant */
     AntStruct *iteration_best_ant;     /* 当前迭代表现最好的蚂蚁 */
     
-    double   **pheromone;               /* pheromone matrix, one entry for each arc */
-    double   **total_info;              /* combination of pheromone and heuristic information */
+    float   **pheromone;               /* pheromone matrix, one entry for each arc */
+    float   **total_info;              /* combination of pheromone and heuristic information */
     
-    double   *prob_of_selection;        /* 依概率选择下一个node */
+    float   *prob_of_selection;        /* 依概率选择下一个node */
     
     int n_ants;                    /* number of ants */
     int nn_ants;                   /* length of nearest neighbor lists for the ants'
                                          solution construction */
     
-    double   max_runtime;               /* maximal allowed run time */
-    double   best_so_far_time;          /* 当前最优解出现的时间 */
+    float   max_runtime;               /* maximal allowed run time */
+    float   best_so_far_time;          /* 当前最优解出现的时间 */
     int best_solution_iter;        /* iteration in which best solution is found */
     
     int rnd_seed;                  /* 用于生成随机数的种子 */
     
-    double last_iter_solution;          /* 上一次迭代的解 */
+    float last_iter_solution;          /* 上一次迭代的解 */
     int iter_stagnate_cnt;         /* 迭代停滞计数器，记录解迭代解停滞的次数 */
     int best_stagnate_cnt;         /* 最优解停滞计数器，最优解未更新的次数 */
     
     /*----- sub problem only -----*/
     int num_subs;                  /* 仅用于parallel版本蚁群, 子问题个数 */
     vector<int> real_nodes;        /* 仅用于sub-problem, sub nodes需要重新编号，因此需要额外数组记录nodes真实的编号 */
-    double   **best_pheromone;          /* 仅用于sub-problem, 记录sub 获得best_so_far_solution时候的信息素
+    float   **best_pheromone;          /* 仅用于sub-problem, 记录sub 获得best_so_far_solution时候的信息素
                                            用于sub 迭代结束时更新至master */
 };
